pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cyberguard"
version = "1.0.0"
description = "Cyber Scam Detection Environment"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "requests",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
